refactor: crear carpetas para mejor organizacion de los componentes
