+++ frontend/src/app/app.routes.ts
@@ -1,10 +1,10 @@
 // app.routes.ts
 import { Routes } from '@angular/router';
-import { Perfil } from './perfil/perfil';
-import { Habitos } from './habitos/habitos';
-import { Registro } from './registro/registro';
-import { NosotrosComponent } from './nosotros/nosotros';
-import { Home } from './home/home'; 
+import { Perfil } from './pages/perfil/perfil';
+import { Habitos } from './pages/habitos/habitos';
+import { Registro } from './pages/registro/registro';
+import { NosotrosComponent } from './pages/nosotros/nosotros';
+import { Home } from './pages/home/home'; 
 
 export const routes: Routes = [
