+++ frontend/src/app/pages/home/call-to-action/call-to-action.css
