+++ frontend/src/app/pages/home/carousel/carousel.css
