+++ frontend/src/app/pages/home/carousel/carousel.html
