+++ frontend/src/app/pages/home/carousel/carousel.spec.ts
