+++ frontend/src/app/pages/home/carousel/carousel.ts
