+++ frontend/src/app/pages/registro/registro.css
