+++ frontend/src/app/pages/registro/registro.html
