+++ frontend/src/app/pages/registro/registro.spec.ts
