+++ frontend/src/app/pages/registro/registro.ts
