+++ frontend/src/app/shared/footer/footer.spec.ts
