+++ frontend/src/app/shared/footer/footer.ts
