+++ frontend/src/app/shared/navbar/navbar.ts
